Add support for embedded tweets in Apple News.
